LFU cache and perfect cache separated into 2 files - lookup_update divided and simplified - removed commented junk - some functions moved to private section
