+++ CMakeLists.txt
@@ -1,7 +1,7 @@
 cmake_minimum_required(VERSION 3.5.0)
-project(LFU_CACHE VERSION 0.1.0 LANGUAGES C CXX)
+project(CACHES VERSION 0.1.0 LANGUAGES C CXX)
 
-add_executable(LFU_CACHE main.cpp caches.hpp settings_parser.hpp)
+add_executable(CACHES main.cpp perfect_cache.hpp LFU_cache.hpp settings_parser.hpp)
 
 set(CPACK_PROJECT_NAME ${PROJECT_NAME})
 set(CPACK_PROJECT_VERSION ${PROJECT_VERSION})
